Refactor + comments
