+++ src/main/resources/lib/paths/public-path.ts
@@ -1,7 +1,8 @@
 import { Content } from '/lib/xp/content';
 import { hasValidCustomPath } from './custom-paths/custom-path-utils';
 import { stripPathPrefix } from './path-utils';
-import { buildLocalePath, isContentLocalized } from '../localization/locale-utils';
+import { isContentLocalized } from '../localization/locale-utils';
+import { buildLocalePath } from './locale-paths';
 
 export const getPublicPath = (content: Content, locale: string): string => {
     const basePath = hasValidCustomPath(content)
